Connected the repositories and view models for movie reviews into the UI. Everything appears to work without error.

+++ app/src/main/java/com/example/bdmi/ui/composables/movie_detail/middle/ReviewHistogram.kt
@@ -27,7 +27,7 @@ import com.example.bdmi.ui.theme.dimens
 
 @Composable
 fun ReviewHistogram(
-    averageRating: Float,
+    averageRating: Double,
     totalReviews: Int,
     ratingCounts: Map<String, Int>,
 ) {
